# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VeriShield AI SaaS API
  description: Enterprise deepfake detection and prevention platform
  version: 1.0.0
  contact:
    name: VeriShield Support
    email: support@verishield.ai
servers:
  - url: https://api.verishield.ai/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
  /analyze/stream:
    post:
      summary: Analyze video/audio stream for deepfakes
      description: Real-time analysis of media streams for deepfake detection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - video_b64
                - audio_b64
                - transaction_amount
              properties:
                video_b64:
                  type: string
                  description: Base64-encoded video data
                audio_b64:
                  type: string
                  description: Base64-encoded audio data
                metadata:
                  type: object
                  description: Additional transaction metadata
                transaction_amount:
                  type: number
                  description: Transaction amount in USD
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision:
                    type: string
                    enum: [REAL, FAKE]
                  reasoning:
                    type: string
                  vision_score:
                    type: number
                    minimum: 0
                    maximum: 1
                  acoustic_score:
                    type: number
                    minimum: 0
                    maximum: 1
                  transaction_amount:
                    type: number
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...